[REFACTOR] refactor transaction test
